Starting to separate out AWS specific implementations in ndbench-aws module.
