+++ ndbench-aws/build.gradle
@@ -1,0 +1,6 @@
+dependencies {
+    compile project(':ndbench-api')
+    compile "com.amazonaws:aws-java-sdk-core:latest.release"
+    testCompile group: 'com.netflix.archaius', name: 'archaius2-guice', version: '2.1.11'
+    testCompile 'com.netflix.governator:governator-test-junit:[1.15.3,)'
+}
