+++ ndbench-aws/src/test/java/com/netflix/ndbench/aws/config/NdbenchAWSCredentialProviderTest.java
@@ -1,0 +1,19 @@
+package com.netflix.ndbench.aws.config;
+
+import com.netflix.archaius.guice.ArchaiusModule;
+import com.netflix.governator.guice.test.ModulesForTesting;
+import com.netflix.governator.guice.test.junit4.GovernatorJunit4ClassRunner;
+import org.junit.Test;
+import org.junit.runner.RunWith;
+
+import javax.inject.Inject;
+
+import static org.junit.Assert.*;
+
+
+public class NdbenchAWSCredentialProviderTest {
+    @Test
+    public void testDefaults() {
+
+    }
+}
