+++ ndbench-dax-plugins/build.gradle
@@ -23,6 +23,7 @@ jacocoTestCoverageVerification {
 }
 
 dependencies {
+    compile project(':ndbench-aws')
     compile project(':ndbench-dynamodb-plugins')
     compile 'com.amazonaws:amazon-dax-client:1.0.200704.0'
     testCompile group: 'com.netflix.archaius', name: 'archaius2-guice', version: '2.1.11'
