+++ ndbench-dax-plugins/src/main/java/com/netflix/ndbench/plugin/dax/configs/DaxModule.java
@@ -21,8 +21,7 @@ import com.google.inject.AbstractModule;
 import com.google.inject.Provides;
 import com.netflix.archaius.ConfigProxyFactory;
 import com.netflix.ndbench.api.plugin.annotations.NdBenchClientPluginGuiceModule;
-import com.netflix.ndbench.plugin.dynamodb.NdbenchAWSCredentialProvider;
-import com.netflix.ndbench.plugin.dynamodb.configs.CredentialsConfiguration;
+import com.netflix.ndbench.aws.config.CredentialsConfiguration;
 
 /**
  * 
@@ -35,7 +34,6 @@ public class DaxModule extends AbstractModule {
 
     @Override
     protected void configure() {
-        bind(AWSCredentialsProvider.class).to(NdbenchAWSCredentialProvider.class);
     }
 
     @Provides
