+++ ndbench-dynamodb-plugins/src/main/java/com/netflix/ndbench/plugin/dynamodb/configs/DynamoDBModule.java
@@ -16,12 +16,11 @@
  */
 package com.netflix.ndbench.plugin.dynamodb.configs;
 
-import com.amazonaws.auth.AWSCredentialsProvider;
 import com.google.inject.AbstractModule;
 import com.google.inject.Provides;
 import com.netflix.archaius.ConfigProxyFactory;
 import com.netflix.ndbench.api.plugin.annotations.NdBenchClientPluginGuiceModule;
-import com.netflix.ndbench.plugin.dynamodb.NdbenchAWSCredentialProvider;
+import com.netflix.ndbench.aws.config.CredentialsConfiguration;
 
 /**
  * 
@@ -34,7 +33,6 @@ public class DynamoDBModule extends AbstractModule {
 
     @Override
     protected void configure() {
-        bind(AWSCredentialsProvider.class).to(NdbenchAWSCredentialProvider.class);
     }
 
     @Provides
