+++ settings.gradle
@@ -12,3 +12,4 @@ include 'ndbench-geode-plugins'
 include 'ndbench-janusgraph-plugins'
 include 'ndbench-sample-plugins'
 include 'ndbench-web'
+include 'ndbench-aws'
